+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Version 0.24
 - improved temporary files handling, don't delete them when linked
 - added tables `datalink.dl_access_web` and `datalink.access_web` table and view 
 - added preliminary `has_web_privilege()` functions
+- remove function `instr()`
 
 Version 0.23
 ------------